Fix retransmission logic, 503 error
Fixes the dispatcher retransmission logic so that the dispatcher
remembers if one of its attempts might have succeeded (even if it never
received a response from the vehicle).

Categorizes an HTTP 503 code as a possible success, since the server may
return this status code if it relays a command to a vehicle but does not
hear back.